refactor: 모달 동작 관련 코드도 분리
